1-2

+++ goods/src/main/java/cn/edu/xmu/oomall/goods/dao/ProductDao.java
@@ -11,6 +11,8 @@ import cn.edu.xmu.oomall.goods.model.po.OnSalePo;
 import cn.edu.xmu.oomall.goods.model.po.OnSalePoExample;
 import cn.edu.xmu.oomall.goods.model.po.ProductDraftPo;
 import cn.edu.xmu.oomall.goods.model.po.ProductPo;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -25,6 +27,7 @@ import static cn.edu.xmu.oomall.core.util.Common.cloneVo;
  **/
 @Repository
 public class ProductDao {
+    private static final Logger logger = LoggerFactory.getLogger(ProductDao.class);
     @Autowired
     private ProductPoMapper productPoMapper;
     @Autowired
@@ -38,24 +41,33 @@ public class ProductDao {
     private Byte prohibit=3;
 
 
-    public ReturnObject publishById(Long id)
+    public ReturnObject publishById(Long shopId,Long id)
     {
         ProductDraftPo productDraftPo;
         productDraftPo=productDraftPoMapper.selectByPrimaryKey(id);
         if(productDraftPo!=null)
         {
+            if(!productDraftPo.getShopId().equals(shopId))
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
             ProductPo productPo=(ProductPo) cloneVo(productDraftPo,ProductPo.class);
             productPo.setState(offshelves);
             try
             {
-                productPoMapper.insert(productPo);
+                int f1=productPoMapper.insert(productPo);
             OnSale onSale=(OnSale)cloneVo(productPo,OnSale.class);
             onSale.setState(OnSale.State.OFFLINE);
-            onSalePoMapper.insert((OnSalePo) cloneVo(onSale,OnSalePo.class));
-            productDraftPoMapper.deleteByPrimaryKey(id);
+            int f2=onSalePoMapper.insert((OnSalePo) cloneVo(onSale,OnSalePo.class));
+            int f3=productDraftPoMapper.deleteByPrimaryKey(id);
+                if(f1==0||f2==0||f3==0)
+                {
+                    return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+                }
             return new ReturnObject((Product)cloneVo(productPo,Product.class));
         }catch(Exception e)
         {
+            logger.error(e.getMessage());
             return new ReturnObject(ReturnNo.INTERNAL_SERVER_ERR,e.getMessage());
         }
         }else
@@ -65,7 +77,7 @@ public class ProductDao {
 
     }
 
-    public ReturnObject onshelvesById(Long id)
+    public ReturnObject onshelvesById(Long shopId,Long id)
     {
         ProductPo productPo;
         try
@@ -75,11 +87,19 @@ public class ProductDao {
             {
                 return new ReturnObject<>(ReturnNo.RESOURCE_ID_NOTEXIST);
             }
+            if(!productPo.getShopId().equals(shopId))
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
         if(productPo.getState().equals(offshelves))
         {
 
             productPo.setState(onshelves);
-            productPoMapper.updateByPrimaryKey(productPo);
+            int f1=productPoMapper.updateByPrimaryKey(productPo);
+            if(f1==0)
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
             List<OnSalePo> onSalePoList;
             OnSalePoExample onSalePoExample=new OnSalePoExample();
             OnSalePoExample.Criteria cr=onSalePoExample.createCriteria();
@@ -92,7 +112,11 @@ public class ProductDao {
             for(OnSalePo onSalePo:onSalePoList)
             {
                 onSalePo.setState(onshelves);
-                onSalePoMapper.updateByPrimaryKey(onSalePo);
+                int f2=onSalePoMapper.updateByPrimaryKey(onSalePo);
+                if(f2==0)
+                {
+                    return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+                }
             }
             return new ReturnObject(ReturnNo.OK);
         }
@@ -102,11 +126,12 @@ public class ProductDao {
         }
         }catch(Exception e)
         {
+            logger.error(e.getMessage());
             return new ReturnObject(ReturnNo.INTERNAL_SERVER_ERR,e.getMessage());
         }
     }
 
-    public ReturnObject offshelvesById(Long id)
+    public ReturnObject offshelvesById(Long shopId,Long id)
     {
         ProductPo productPo;
         try
@@ -116,10 +141,18 @@ public class ProductDao {
             {
                 return new ReturnObject<>(ReturnNo.RESOURCE_ID_NOTEXIST);
             }
+            if(!productPo.getShopId().equals(shopId))
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
         if(productPo.getState().equals(onshelves))
         {
             productPo.setState(offshelves);
-            productPoMapper.updateByPrimaryKey(productPo);
+            int f1=productPoMapper.updateByPrimaryKey(productPo);
+            if(f1==0)
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
             List<OnSalePo> onSalePoList;
             OnSalePoExample onSalePoExample=new OnSalePoExample();
             OnSalePoExample.Criteria cr=onSalePoExample.createCriteria();
@@ -132,7 +165,11 @@ public class ProductDao {
             for(OnSalePo onSalePo:onSalePoList)
             {
                 onSalePo.setState(offshelves);
-                onSalePoMapper.updateByPrimaryKey(onSalePo);
+                int f2=onSalePoMapper.updateByPrimaryKey(onSalePo);
+                if(f2==0)
+                {
+                    return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+                }
             }
             return new ReturnObject(ReturnNo.OK);
         }
@@ -142,12 +179,13 @@ public class ProductDao {
         }
     }catch (Exception e)
     {
+        logger.error(e.getMessage());
         return new ReturnObject(ReturnNo.INTERNAL_SERVER_ERR,e.getMessage());
     }
 
     }
 
-    public ReturnObject allowProductById(Long id)
+    public ReturnObject allowProductById(Long shopId,Long id)
     {
         ProductPo productPo;
         try
@@ -156,6 +194,10 @@ public class ProductDao {
             if(productPo==null)
             {
                 return new ReturnObject<>(ReturnNo.RESOURCE_ID_NOTEXIST);
+            }
+            if(!productPo.getShopId().equals(shopId))
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
             }
         if(productPo.getState().equals(prohibit))
         {
@@ -176,7 +218,11 @@ public class ProductDao {
                 if(onSalePo.getState().equals(prohibit))
                 {
                     onSalePo.setState(offshelves);
-                    onSalePoMapper.updateByPrimaryKey(onSalePo);
+                    int f1=onSalePoMapper.updateByPrimaryKey(onSalePo);
+                    if(f1==0)
+                    {
+                        return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+                    }
                 }
                 else
                 {
@@ -191,11 +237,12 @@ public class ProductDao {
         }
     }catch (Exception e)
     {
+        logger.error(e.getMessage());
         return new ReturnObject(ReturnNo.INTERNAL_SERVER_ERR,e.getMessage());
     }
     }
 
-    public ReturnObject prohibitProductById(Long id)
+    public ReturnObject prohibitProductById(Long shopId,Long id)
     {
         ProductPo productPo;
         try
@@ -205,10 +252,18 @@ public class ProductDao {
             {
                 return new ReturnObject<>(ReturnNo.RESOURCE_ID_NOTEXIST);
             }
+            if(!productPo.getShopId().equals(shopId))
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
         if(productPo.getState().equals(onshelves)||productPo.getState().equals(offshelves))
         {
             productPo.setState(prohibit);
-            productPoMapper.updateByPrimaryKey(productPo);
+            int f1=productPoMapper.updateByPrimaryKey(productPo);
+            if(f1==0)
+            {
+                return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+            }
             List<OnSalePo> onSalePoList;
             OnSalePoExample onSalePoExample=new OnSalePoExample();
             OnSalePoExample.Criteria cr=onSalePoExample.createCriteria();
@@ -219,7 +274,11 @@ public class ProductDao {
                 if(onSalePo.getState().equals(onshelves)||onSalePo.getState().equals(offshelves))
                 {
                     onSalePo.setState(prohibit);
-                    onSalePoMapper.updateByPrimaryKey(onSalePo);
+                    int f2=onSalePoMapper.updateByPrimaryKey(onSalePo);
+                    if(f2==0)
+                    {
+                        return new ReturnObject(ReturnNo.FIELD_NOTVALID);
+                    }
                 }
                 else
                 {
@@ -234,9 +293,8 @@ public class ProductDao {
         }
     }catch (Exception e)
     {
+        logger.error(e.getMessage());
         return new ReturnObject(ReturnNo.INTERNAL_SERVER_ERR,e.getMessage());
     }
     }
-
-
 }
